[#14] Fixed and improved tests

+++ project2/tests/0-mult.cc
@@ -1,8 +1,8 @@
 int main() {
     int a = 3 * 4;
-    int b = 3.0 * 7.0;
-    int c = 3.0 * 4;
-    int d = 4 * 3.0;
+    double b = 3.0 * 7.0;
+    double c = 3.0 * 4;
+    double d = 4 * 3.0;
     
     int x = 3 * true;
     int y = true * false;
